boilerplate/requirements/base.txt
# Flask and Extensions
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Werkzeug==2.2.3

# Database
SQLAlchemy==2.0.28
alembic==1.13.1
PyMySQL==1.0.2
mysql-connector-python==8.0.33

# Utilities
python-dotenv==1.0.0
email_validator==2.0.0
requests==2.31.0
python-slugify==8.0.1 